+++ lib/td_lm_web/controllers/relation_controller.ex
@@ -181,7 +181,8 @@ defmodule TdLmWeb.RelationController do
     end
   end
 
-  defp refresh_attributes(relation, relation_side, relation_id_key, target_type) do
+  defp refresh_attributes(relation, relation_side, relation_id_key, target_type)
+       when target_type in ["business_concept", "ingest"] do
     relation_side_attrs =
       relation
       |> Map.get(:context)
@@ -192,44 +193,54 @@ defmodule TdLmWeb.RelationController do
         relation
 
       relation_side_attrs ->
-        version_id = get_version_id(target_type, Map.get(relation, relation_id_key))
-        name = get_name(target_type, Map.get(relation, relation_id_key))
+        cached = fetch_attributes(Map.get(relation, relation_id_key), target_type)
+        version_id = Map.get(cached, :version_id)
+        name = Map.get(cached, :name)
 
         relation_side_attrs =
           relation_side_attrs
-          |> Map.put("version_id", version_id)
           |> Map.put("name", name)
+          |> Map.put("version_id", version_id)
 
-        put_attrs_in_context(relation, relation_side, relation_side_attrs, version_id)
+        context = Map.put(relation.context, relation_side, relation_side_attrs)
+        Map.put(relation, :context, context)
     end
   end
 
-  defp put_attrs_in_context(relation, _relation_side, _relation_side_attrs, nil) do
-    relation
+  defp refresh_attributes(relation, _, _, _), do: relation
+
+  defp fetch_attributes(entity_id, "business_concept") do
+    case ConceptCache.get(entity_id) do
+      {:ok, concept = %{}} ->
+        concept
+        |> Map.take([:name, :business_concept_version_id])
+        |> Enum.map(fn
+          {:business_concept_version_id, version} -> {:version_id, version}
+          other -> other
+        end)
+        |> Enum.into(%{})
+
+      _ ->
+        %{}
+    end
   end
 
-  defp put_attrs_in_context(relation, relation_side, relation_side_attrs, _version_id) do
-    context = Map.put(relation.context, relation_side, relation_side_attrs)
-    Map.put(relation, :context, context)
+  defp fetch_attributes(entity_id, "ingest") do
+    case IngestCache.get(entity_id) do
+      {:ok, ingest = %{}} ->
+        ingest
+        |> Map.take([:name, :ingest_version_id])
+        |> Enum.map(fn
+          {:ingest_version_id, version} -> {:version_id, version}
+          other -> other
+        end)
+        |> Enum.into(%{})
+
+      _ ->
+        %{}
+    end
   end
 
-  defp get_version_id("business_concept", entity_id) do
-    {:ok, id} = ConceptCache.get(entity_id, :business_concept_version_id)
-    id
-  end
+  defp fetch_attributes(_entity_id, _target_type), do: %{}
 
-  defp get_version_id("ingest", entity_id) do
-    IngestCache.get_ingest_version_id(entity_id)
-  end
-
-  defp get_version_id(_, _) do
-  end
-
-  defp get_name("business_concept", entity_id) do
-    {:ok, name} = ConceptCache.get(entity_id, :name)
-    name
-  end
-
-  defp get_name(_, _) do
-  end
 end